chapter_005/src/test/java/ru/job4j/list/CustomQueueTest.java
package ru.job4j.list;

import org.junit.Test;

import static org.hamcrest.MatcherAssert.assertThat;
import static  org.hamcrest.Matchers.is;

public class CustomQueueTest {
    @Test
    public void pushPollTests() {
        CustomQueue<Integer> queue = new CustomQueue<>();
        queue.push(1);
        queue.push(2);
        queue.push(3);
        assertThat(queue.poll(), is(1));
        assertThat(queue.poll(), is(2));
        assertThat(queue.poll(), is(3));
        assertThat(queue.poll() == null, is(true));
        queue.push(1);
        queue.push(2);
        assertThat(queue.poll(), is(1));
        queue.push(3);
        assertThat(queue.poll(), is(2));
        assertThat(queue.poll(), is(3));
        assertThat(queue.poll() == null, is(true));
        assertThat(queue.poll() == null, is(true));
        queue.push(1);
        assertThat(queue.poll(), is(1));
    }
}
